MintableToken: Replaced canMint() with ifMinting()
According to https://medium.com/@maraoz/f1d742b94756:
"The name of the canMint modifier might imply that it succeeds
if the transaction can mint tokens. However, it doesn’t check that
the sender is a mint agent, which is checked separately by a different
modifier. Consider renaming the modifier."

Selected ifMinting(), since it throws if minting is finished.