PricingStrategy: Implementing isSane() as abstract
According to https://medium.com/@maraoz/f1d742b94756:
"isSane function in line 17 of PricingStrategy.sol could be made abstract.
This would make it harder for a developer implementing a child contract
to forget implementing the method."

Also modified FlatPricing, which was the only pricing strategy not to
implement its own version of isSane().