deploycontracts: Added sanitycheck for the git repository
Now, if the script sees that the repository is not clean (not committed),
it does not proceed. Also, it will abort on non-master branches (by default,
possible to skip by using ICO_DEPLOY_ANYWAY env), and also it will abort if
you haven't pulled for a week.

The deployment script is also adding the commit hash to the source verified
by Etherscan.

Also added crowdsales/*.py to .gitignore for easier tracking of changes.